+++ src/tinyEmmiter.js
